add types to type file, rename tabletitle, remove edit buttom temp.
